orb with cuda working:

+++ src/services/countLeftGPUService.h
@@ -5,7 +5,7 @@ class ServiceCountLeftGPU : public TraverseCombinePST {
 public:
     static constexpr int max_cells = 8192;
     typedef struct Cell input; // Array of Cells
-    typedef uint64_t output;   // Array of counts
+    typedef uint output;   // Array of counts
     explicit ServiceCountLeftGPU(PST pst)
         : TraverseCombinePST(pst,PST_COUNTLEFTGPU,max_cells*sizeof(input),max_cells*sizeof(output),"CountLeftGPU") {}
 protected:
